+++ internal/extension/proc/rpc_timeout_test.go
@@ -1,4 +1,4 @@
-package extension
+package proc
 
 import (
 	"context"